+++ CHANGELOG.md
@@ -11,6 +11,9 @@ Versions since `1.7.0` only track ABI breaks and not API breaks.
 - Code Page 437: Character 0x7F is now assigned to 0x2302 (HOUSE).
   [#134](https://github.com/libtcod/libtcod/pull/134)
 
+### Deprecated
+- All color constants have been deprecated.
+
 ## [1.23.0] - 2022-10-22
 ### Added
 - Added new C print functions: ``TCOD_printf_rgb``, ``TCOD_printn_rgb``, and ``TCOD_vprintf_rgb``.